requirements.txt
flask==3.0.3
flask-session==0.8.0
python-dotenv==1.0.1
langchain==0.2.14
langchain-google-genai==1.0.7
gunicorn==22.0.0
